+++ libs/ui/src/lib/tunnel/index.ts
@@ -2,3 +2,4 @@ export * from './tunnel.guard';
 export * from './tunnel.service';
 export * from './page/page.module';
 export * from './layout/layout.module';
+export * from './tunnel.model';
